Modified Del3_game
